+++ src/components/sidebar/sidebar-toggle.tsx
@@ -17,7 +17,8 @@ export function SidebarToggle({
           className={cn("h-8 px-2 md:h-fit md:px-2", className)}
           data-testid="sidebar-toggle-button"
           onClick={toggleSidebar}
-          variant="outline">
+          variant="outline"
+        >
           <SidebarLeftIcon size={16} />
         </Button>
       </TooltipTrigger>