+++ src/main.tsx
@@ -1,13 +1,28 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { MiniKitProvider } from "@worldcoin/minikit-js/minikit-provider";
+import { isWorldApp } from "@/lib/utils";
 import App from "./App";
 import "./index.css";
 
-createRoot(document.getElementById("root")!).render(
+const rootElement = document.getElementById("root");
+if (!rootElement) {
+  throw new Error("Root element not found");
+}
+
+const AppWithProviders = () => {
+  if (isWorldApp()) {
+    return (
+      <MiniKitProvider>
+        <App />
+      </MiniKitProvider>
+    );
+  }
+  return <App />;
+};
+
+createRoot(rootElement).render(
   <StrictMode>
-    <MiniKitProvider>
-      <App />
-    </MiniKitProvider>
+    <AppWithProviders />
   </StrictMode>,
 );